started interrupt at_cmd
