+++ .gitignore
@@ -1,3 +1,5 @@
 .pio
 .vscode
 logs
+__pycache__
+.DS_Store
