+++ scripts/config.py
@@ -1,91 +1,33 @@
 #!/bin/python3
 """
-When run in main, the script will record and send over the user's 
-desired communication configurations to a connected Arduino for RF 
+When run in main, the script will record and send over the user's
+desired communication configurations to a connected Arduino for RF
 file transfers.
 
-Params: 
+Params:
     sys.argv[1]:
         Absolute path of Serial port
 
-    sys.argv[2]: 
+    sys.argv[2]:
         Baudrate of Serial port
 
 Sends:
-    start_signal: 
+    start_signal:
         signal to the Arduino that config params are about to be sent
         this step is important for flushing the Arduino's Serial before
         configurations are sent
 
-    channel: 
+    channel:
         0-125
-        
-    address: 
+
+    address:
         0-10000
 """
 
 import sys
 import serial
-
-
-# We are going to store the configured integers in our Arduino in a Union, 
-# by sending over our individal bytes and storing them in memory.  
-# Thus, we need to consider the endianess of our Arduino
-ENDIANESS = 'little'
-
-# byte to send to Arduino to signal that we are about to send a file
-HANDSHAKE_CHAR_BYTE = 9
-
-# How many consecutive signals we need to send over 
-HANDSHAKE_REPS = 5
-
-# RF frequency channel bounds
-MAX_CHANNEL = 125
-MIN_CHANNEL = 0
-
-# Data pipe address bounds
-MAX_ADDRESS = 10000
-MIN_ADDRESS = 0
-
-# All numbers 0-9 as strings, used for input error handling
-NUMBERS = set([str(x) for x in range(10)])
-
-
-def sendFlushSerialSignal(ser):
-    """
-    Sends to the Arduino HANDSHAKE_BYTE HANDSHAKE_REPS times.  On the Arduino side, this
-    function relies on the flushSerial() function in the serial_io class to flush the
-    serial and prepare the Arsuino for communication.
-
-    Params:
-        ser:
-            Our initiallized pyserial serial port
-
-    Outputs:
-        None
-    """
-    # signal to the Arduino that we are about to send over configuration data
-    flush_signal = bytearray()
-    flush_signal.extend([HANDSHAKE_CHAR_BYTE for _ in range(HANDSHAKE_REPS)])
-    ser.write(flush_signal)
-
-
-def getIntInput():
-    """
-    Returns -1 if our input is not an integer, and the input if it is an integer.
-
-    Params:
-        None
-    
-    Outputs:
-        int
-    """
-    int_input = input()
-    if not int_input or not all([i in NUMBERS for i in int_input]):
-        int_input = -1
-
-    return int(int_input)
-
+from arduino_serial_io import (MIN_CHANNEL, MAX_CHANNEL, MIN_ADDRESS,
+                               MAX_ADDRESS, ENDIANESS, getIntInput, flushSerial)
 
 if __name__ == "__main__":
 
@@ -94,15 +36,15 @@ if __name__ == "__main__":
     address = -1
 
     print("Enter channel ({0}-{1}): ".format(MIN_CHANNEL, MAX_CHANNEL))
-    while not (MIN_CHANNEL <= channel <= MAX_CHANNEL):
+    while not MIN_CHANNEL <= channel <= MAX_CHANNEL:
         channel = getIntInput()
-        if not (MIN_CHANNEL <= channel <= MAX_CHANNEL):
+        if not MIN_CHANNEL <= channel <= MAX_CHANNEL:
             print("Invalid input. Re-enter channel ({0}-{1}): ".format(MIN_CHANNEL, MAX_CHANNEL))
-    
+
     print("Enter address ({0}-{1}): ".format(MIN_ADDRESS, MAX_ADDRESS))
-    while not (MIN_ADDRESS <= address <= MAX_ADDRESS):
+    while not MIN_ADDRESS <= address <= MAX_ADDRESS:
         address = getIntInput()
-        if not (MIN_ADDRESS <= address <= MAX_ADDRESS):
+        if not MIN_ADDRESS <= address <= MAX_ADDRESS:
             print("Invalid input. Re-enter address ({0}-{1}): ".format(MIN_ADDRESS, MAX_ADDRESS))
 
     # configuring our serial
@@ -110,12 +52,12 @@ if __name__ == "__main__":
     ser.port = sys.argv[1]
     ser.baudrate = int(sys.argv[2])
     ser.open()
-    sendFlushSerialSignal(ser)
-    
+    flushSerial(ser)
+
     # sending over our configurations
     channel = channel.to_bytes(1, byteorder=ENDIANESS)
-    address = address.to_bytes(4, byteorder=ENDIANESS)    
+    address = address.to_bytes(4, byteorder=ENDIANESS)
     ser.write(channel)
     ser.write(address)
-    
+
     ser.close()
